Added Default and Sensitive elements for nginx and spark
Thought:
Is there a better way to connect the sensitive elements and the dealing funcs than the way we currently do it?
The need for speciric dealing functions will be very different between applications.
Maybe if we just linked them directly in the sensitive list as a key value pair.
Thoughts?